requirements.txt
click==8.1.7
pyyaml==6.0.1
boto3==1.29.6
python-dotenv==1.0.0
requests==2.31.0
rich==13.6.0 